+++ python-project/react-app/src/components/auth/SignUpForm/SignUpForm.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { useSelector, useDispatch } from 'react-redux'
 import { Redirect } from 'react-router-dom';
-import { signUp } from '../../store/session';
+import { signUp } from '../../../store/session';
 import './SignUpForm.css'
 
 const SignUpForm = () => {
